+++ src/main/java/com/magneticraft2/common/block/BlockMagneticraft2.java
@@ -31,7 +31,7 @@ public abstract class BlockMagneticraft2 extends Block implements TOPDriver {
         if (te instanceof TileEntityMagneticraft2){
             TileEntityMagneticraft2 tile = (TileEntityMagneticraft2) te;
             te.getCapability(CapabilityHeat.HEAT).ifPresent(h -> {
-                probeInfo.horizontal(probeInfo.defaultLayoutStyle()).progress(h.getHeatStored() + 0 % 100, 100, probeInfo.defaultProgressStyle().suffix(" HEAT").borderColor(0xFF555555));
+                probeInfo.horizontal(probeInfo.defaultLayoutStyle()).progress((int) (h.getHeatStored() + 0 % 100), 100, probeInfo.defaultProgressStyle().suffix(" HEAT").borderColor(0xFF555555));
             });
             te.getCapability(CapabilityWatt.WATT).ifPresent(h -> {
                 probeInfo.horizontal(probeInfo.defaultLayoutStyle()).progress(h.getWattStored() + 0 % 100, 100, probeInfo.defaultProgressStyle().suffix(" WATT").borderColor(0xFF555555));
